requirements.txt
Flask==2.2.3
python-dotenv==1.0.0
stripe==5.3.0